requirements.txt
qq-botpy==1.1.2
openai==1.3.2
